+++ src/automation/azext_automation/vendored_sdks/automation/models/_automation_client_enums.py
@@ -6,27 +6,18 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from enum import Enum, EnumMeta
-from six import with_metaclass
-
-class _CaseInsensitiveEnumMeta(EnumMeta):
-    def __getitem__(self, name):
-        return super().__getitem__(name.upper())
-
-    def __getattr__(cls, name):
-        """Return the enum member matching `name`
-        We use __getattr__ instead of descriptors or inserting into the enum
-        class' __dict__ in order to support `name` and `value` being both
-        properties for enum members (which live in the class' __dict__) and
-        enum members themselves.
-        """
-        try:
-            return cls._member_map_[name.upper()]
-        except KeyError:
-            raise AttributeError(name)
+from enum import Enum
+from azure.core import CaseInsensitiveEnumMeta
 
 
-class AutomationAccountState(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
+class AgentRegistrationKeyName(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """Gets or sets the agent registration key name - primary or secondary.
+    """
+
+    PRIMARY = "primary"
+    SECONDARY = "secondary"
+
+class AutomationAccountState(str, Enum, metaclass=CaseInsensitiveEnumMeta):
     """Gets status of account.
     """
 
@@ -34,21 +25,71 @@ class AutomationAccountState(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)
     UNAVAILABLE = "Unavailable"
     SUSPENDED = "Suspended"
 
-class AutomationKeyName(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
+class AutomationKeyName(str, Enum, metaclass=CaseInsensitiveEnumMeta):
     """Automation key name.
     """
 
     PRIMARY = "Primary"
     SECONDARY = "Secondary"
 
-class AutomationKeyPermissions(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
+class AutomationKeyPermissions(str, Enum, metaclass=CaseInsensitiveEnumMeta):
     """Automation key permissions.
     """
 
     READ = "Read"
     FULL = "Full"
 
-class HttpStatusCode(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
+class ContentSourceType(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """Gets or sets the content source type.
+    """
+
+    EMBEDDED_CONTENT = "embeddedContent"
+    URI = "uri"
+
+class CountType(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+
+    STATUS = "status"
+    NODECONFIGURATION = "nodeconfiguration"
+
+class CreatedByType(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """The type of identity that created the resource.
+    """
+
+    USER = "User"
+    APPLICATION = "Application"
+    MANAGED_IDENTITY = "ManagedIdentity"
+    KEY = "Key"
+
+class DscConfigurationState(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """Gets or sets the state of the configuration.
+    """
+
+    NEW = "New"
+    EDIT = "Edit"
+    PUBLISHED = "Published"
+
+class EncryptionKeySourceType(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """Encryption Key Source
+    """
+
+    MICROSOFT_AUTOMATION = "Microsoft.Automation"
+    MICROSOFT_KEYVAULT = "Microsoft.Keyvault"
+
+class GraphRunbookType(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """Runbook Type
+    """
+
+    GRAPH_POWER_SHELL = "GraphPowerShell"
+    GRAPH_POWER_SHELL_WORKFLOW = "GraphPowerShellWorkflow"
+
+class GroupTypeEnum(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """Type of the HybridWorkerGroup.
+    """
+
+    USER = "User"
+    SYSTEM = "System"
+
+class HttpStatusCode(str, Enum, metaclass=CaseInsensitiveEnumMeta):
 
     CONTINUE_ENUM = "Continue"
     SWITCHING_PROTOCOLS = "SwitchingProtocols"
@@ -98,7 +139,7 @@ class HttpStatusCode(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
     GATEWAY_TIMEOUT = "GatewayTimeout"
     HTTP_VERSION_NOT_SUPPORTED = "HttpVersionNotSupported"
 
-class JobProvisioningState(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
+class JobProvisioningState(str, Enum, metaclass=CaseInsensitiveEnumMeta):
     """The provisioning state of the resource.
     """
 
@@ -107,7 +148,7 @@ class JobProvisioningState(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
     SUSPENDED = "Suspended"
     PROCESSING = "Processing"
 
-class JobStatus(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
+class JobStatus(str, Enum, metaclass=CaseInsensitiveEnumMeta):
     """Gets or sets the status of the job.
     """
 
@@ -125,7 +166,7 @@ class JobStatus(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
     RESUMING = "Resuming"
     REMOVING = "Removing"
 
-class JobStreamType(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
+class JobStreamType(str, Enum, metaclass=CaseInsensitiveEnumMeta):
     """Gets or sets the stream type.
     """
 
@@ -137,7 +178,61 @@ class JobStreamType(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
     VERBOSE = "Verbose"
     ANY = "Any"
 
-class RunbookState(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
+class LinuxUpdateClasses(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """Update classifications included in the software update configuration.
+    """
+
+    UNCLASSIFIED = "Unclassified"
+    CRITICAL = "Critical"
+    SECURITY = "Security"
+    OTHER = "Other"
+
+class ModuleProvisioningState(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """Gets or sets the provisioning state of the module.
+    """
+
+    CREATED = "Created"
+    CREATING = "Creating"
+    STARTING_IMPORT_MODULE_RUNBOOK = "StartingImportModuleRunbook"
+    RUNNING_IMPORT_MODULE_RUNBOOK = "RunningImportModuleRunbook"
+    CONTENT_RETRIEVED = "ContentRetrieved"
+    CONTENT_DOWNLOADED = "ContentDownloaded"
+    CONTENT_VALIDATED = "ContentValidated"
+    CONNECTION_TYPE_IMPORTED = "ConnectionTypeImported"
+    CONTENT_STORED = "ContentStored"
+    MODULE_DATA_STORED = "ModuleDataStored"
+    ACTIVITIES_STORED = "ActivitiesStored"
+    MODULE_IMPORT_RUNBOOK_COMPLETE = "ModuleImportRunbookComplete"
+    SUCCEEDED = "Succeeded"
+    FAILED = "Failed"
+    CANCELLED = "Cancelled"
+    UPDATING = "Updating"
+
+class OperatingSystemType(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """Target operating system for the software update configuration.
+    """
+
+    WINDOWS = "Windows"
+    LINUX = "Linux"
+
+class ProvisioningState(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """The provisioning state of the job.
+    """
+
+    COMPLETED = "Completed"
+    FAILED = "Failed"
+    RUNNING = "Running"
+
+class ResourceIdentityType(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """The identity type.
+    """
+
+    SYSTEM_ASSIGNED = "SystemAssigned"
+    USER_ASSIGNED = "UserAssigned"
+    SYSTEM_ASSIGNED_USER_ASSIGNED = "SystemAssigned, UserAssigned"
+    NONE = "None"
+
+class RunbookState(str, Enum, metaclass=CaseInsensitiveEnumMeta):
     """Gets or sets the state of the runbook.
     """
 
@@ -145,7 +240,7 @@ class RunbookState(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
     EDIT = "Edit"
     PUBLISHED = "Published"
 
-class RunbookTypeEnum(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
+class RunbookTypeEnum(str, Enum, metaclass=CaseInsensitiveEnumMeta):
     """Gets or sets the type of the runbook.
     """
 
@@ -155,10 +250,95 @@ class RunbookTypeEnum(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
     POWER_SHELL = "PowerShell"
     GRAPH_POWER_SHELL_WORKFLOW = "GraphPowerShellWorkflow"
     GRAPH_POWER_SHELL = "GraphPowerShell"
+    PYTHON2 = "Python2"
+    PYTHON3 = "Python3"
 
-class SkuNameEnum(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
+class ScheduleDay(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """Day of the occurrence. Must be one of monday, tuesday, wednesday, thursday, friday, saturday,
+    sunday.
+    """
+
+    MONDAY = "Monday"
+    TUESDAY = "Tuesday"
+    WEDNESDAY = "Wednesday"
+    THURSDAY = "Thursday"
+    FRIDAY = "Friday"
+    SATURDAY = "Saturday"
+    SUNDAY = "Sunday"
+
+class ScheduleFrequency(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """Gets or sets the frequency of the schedule.
+    """
+
+    ONE_TIME = "OneTime"
+    DAY = "Day"
+    HOUR = "Hour"
+    WEEK = "Week"
+    MONTH = "Month"
+    #: The minimum allowed interval for Minute schedules is 15 minutes.
+    MINUTE = "Minute"
+
+class SkuNameEnum(str, Enum, metaclass=CaseInsensitiveEnumMeta):
     """Gets or sets the SKU name of the account.
     """
 
     FREE = "Free"
     BASIC = "Basic"
+
+class SourceType(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """The source type. Must be one of VsoGit, VsoTfvc, GitHub.
+    """
+
+    VSO_GIT = "VsoGit"
+    VSO_TFVC = "VsoTfvc"
+    GIT_HUB = "GitHub"
+
+class StreamType(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """The type of the sync job stream.
+    """
+
+    ERROR = "Error"
+    OUTPUT = "Output"
+
+class SyncType(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """The sync type.
+    """
+
+    PARTIAL_SYNC = "PartialSync"
+    FULL_SYNC = "FullSync"
+
+class TagOperators(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """Filter VMs by Any or All specified tags.
+    """
+
+    ALL = "All"
+    ANY = "Any"
+
+class TokenType(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """The token type. Must be either PersonalAccessToken or Oauth.
+    """
+
+    PERSONAL_ACCESS_TOKEN = "PersonalAccessToken"
+    OAUTH = "Oauth"
+
+class WindowsUpdateClasses(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """Update classification included in the software update configuration. A comma separated string
+    with required values
+    """
+
+    UNCLASSIFIED = "Unclassified"
+    CRITICAL = "Critical"
+    SECURITY = "Security"
+    UPDATE_ROLLUP = "UpdateRollup"
+    FEATURE_PACK = "FeaturePack"
+    SERVICE_PACK = "ServicePack"
+    DEFINITION = "Definition"
+    TOOLS = "Tools"
+    UPDATES = "Updates"
+
+class WorkerType(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """Type of the HybridWorker.
+    """
+
+    HYBRID_V1 = "HybridV1"
+    HYBRID_V2 = "HybridV2"